requirements.txt
spacy
bert-extractive-summarizer
langdetect
deep-translator
nltk
torch
sacremoses

